Move ValueMetaGeometry to its own module. Package separately from plugins. Add geometry-datatype as module dependency on pentaho-gis-plugins module. Edit gitignore to ignore idea specific files.
