+++ .gitignore
@@ -2,3 +2,5 @@ target
 .settings
 .classpath
 .project
+.idea/
+*.iml
