+++ pom.xml
@@ -21,5 +21,6 @@
     <module>mifmid</module>
     <module>shp</module>
     <module>spatialite</module>
+    <module>geometry-datatype</module>
   </modules>
 </project>
